+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added game repairing function
 - Added french translations (#81)
+- Used `open` library for xdg (#84)
+- Added dynamic components loading
 
 ### Fixed
 
 - Forced `format_lang` to return regions for language codes
-- Fixed titlebar behavior on some DEs 
+- Fixed titlebar behavior on some DEs
 
 ### Changed